refactor(api): Cache Spotify API data server-side
Move data fetching logic from local storage to server-side caching by migrating hooks files to the /app/api/spotify folder.

Changes Made:
- Migrated hooks files responsible for fetching data from Spotify API.
- Data that was previously stored in the browser's local storage is now cached on the server side.

Reasoning:
- By caching the fetched data on the server side, we reduce the frequency of API calls, mitigating the risk of hitting usage limits imposed by the Spotify API.
