+++ src/app/api/spotify/playlistData/index.tsx
@@ -1,14 +1,14 @@
-import fs from 'fs/promises';
-import path from 'path';
-import GetNewPlaylistData from '../newPlaylistData/route';
+import fs from "fs/promises";
+import path from "path";
+import GetNewPlaylistData from "../newPlaylistData";
 
 const onRepeatPlaylistId = process.env.NEXT_PUBLIC_ON_REPEAT_PLAYLIST_ID;
-const cacheFilePath = path.resolve('.cache/cachedPlaylistData.json');
+const cacheFilePath = path.resolve(".cache/cachedPlaylistData.json");
 
 export default async function GetPlaylistData() {
   try {
     // Check if the cache file exists
-    const cacheFileContent = await fs.readFile(cacheFilePath, 'utf-8');
+    const cacheFileContent = await fs.readFile(cacheFilePath, "utf-8");
     const cachedPlaylistData = JSON.parse(cacheFileContent);
 
     const currentTime = new Date().getTime();
@@ -17,7 +17,7 @@ export default async function GetPlaylistData() {
       const timeDifference = cachedPlaylistData.expirationDate - currentTime;
       // Convert milliseconds to days
       const daysDifference = Math.ceil(timeDifference / (1000 * 60 * 60 * 24));
-      console.log('use cached data,', `expired in ${daysDifference} days`);
+      console.log("use cached data,", `expired in ${daysDifference} days`);
 
       return cachedPlaylistData.data;
     }
@@ -25,7 +25,7 @@ export default async function GetPlaylistData() {
     // File doesn't exist or error reading file, proceed to fetch new data
   }
 
-  console.log('use new data');
+  console.log("use new data");
   try {
     const result = await GetNewPlaylistData(onRepeatPlaylistId as string);
     const expirationDate = new Date().getTime() + 7 * 24 * 60 * 60 * 1000;
@@ -35,12 +35,12 @@ export default async function GetPlaylistData() {
     await fs.writeFile(
       cacheFilePath,
       JSON.stringify({ data: result, expirationDate }),
-      'utf-8'
+      "utf-8"
     );
 
-    console.log('caching new data');
+    console.log("caching new data");
     return result;
   } catch (error) {
-    console.error('Error fetching data:', error);
+    console.error("Error fetching data:", error);
   }
 }
